+++ app/src/main/java/ru/ramazan/tubesleeper/MainActivity.java
@@ -41,39 +41,6 @@ public class MainActivity extends AppCompatActivity {
         if (!Objects.equals(youtubeUrl, ""))
             new VideoLinkProvider(webView).setVideoLink(youtubeUrl);
 
-        //if (!Objects.equals(VideoLink.videoLink, ""))
-        //     webView.loadUrl(VideoLink.videoLink); // videoLink Содержит ссылку на видео. То есть, если её открыть, то уже откроется чистое видео.
-        //    webView.loadUrl("https://vk.com"); // Однако пока не получается попросить WebView открыть эту ссылку. Надо понять, как попрость webView открыть нашу ссылку.
-        // Тут я ради эксперимента попробовал открыть вконтакте. Но как видел, не работает. Разберись короче, как перейти на другую ссылку через один экземпрял WebView.
-        // Там в группах поспрашивай и в гугле поиши.
-        // Всё давай, завтра утром жду рабочий вариант.
-            //webView.evaluateJavascript("(function() {alert('сообщение');})();", s -> {
-            ////    Log.d("LogName", s); // Prints: "this"
-            //}); // Посмотри что делает эта функция. Возможно это иньекция пользовательского скрипта на страницу.
-
-
-
-
-
-        /*
-
-        WebViewClient webViewClient = new WebViewClient() {
-
-            public boolean shouldOverrideUrlLoading(WebView view, String url) {
-                view.loadUrl(url);
-                return true;
-            }
-
-
-            public boolean shouldOverrideUrlLoading(WebView view, WebResourceRequest request) {
-                view.loadUrl(request.getUrl().toString());
-                return true;
-            }
-        };
-
-         */
-
-
     }
