+++ app/src/main/java/ru/ramazan/tubesleeper/VideoActivity.java
@@ -6,6 +6,7 @@ import android.annotation.SuppressLint;
 import android.os.Bundle;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
+import android.webkit.WebViewClient;
 
 public class VideoActivity extends AppCompatActivity {
 
@@ -15,12 +16,17 @@ public class VideoActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.video_activity);
 
-        @SuppressLint({"MissingInflatedId", "LocalSuppress"})
         WebView webView = findViewById(R.id.videoWebView);
         WebSettings webSetting = webView.getSettings();
         webSetting.setJavaScriptEnabled(true);
 
         webView.loadUrl(VideoLink.videoLink);
+        webView.setWebViewClient(new WebViewClient(){
+            @Override
+            public void onPageFinished(WebView webView, String url) {
+                super.onPageFinished(webView, url);
+            }
+        });
 
     }
 }
