+++ app/src/main/java/ru/ramazan/tubesleeper/VideoLinkProvider.java
@@ -1,18 +1,10 @@
 package ru.ramazan.tubesleeper;
 
-import android.annotation.SuppressLint;
-import android.app.Activity;
-import android.content.Intent;
-import android.util.Log;
-import android.webkit.JavascriptInterface;
-import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
 
 public class VideoLinkProvider {
-
-    VideoLinkProvider(){}
 
     VideoLinkProvider(WebView webView) {
         this.webView = webView;
@@ -33,24 +25,12 @@ public class VideoLinkProvider {
                 injectJavaScript();
             }
         });
-        webView.addJavascriptInterface(new JSBridge(), "Bridge");
+        webView.addJavascriptInterface(new VideoOpener(), "VideoOpener");
 
     }
 
     public void injectJavaScript() {
-        webView.loadUrl("javascript:(function(){setTimeout(() => { let url = document.querySelector('.download-icon').href; Bridge.calledFromJS(url) }, 10000) })()");
-    }
-
-    static class JSBridge {
-        @SuppressLint("SetJavaScriptEnabled")
-        @JavascriptInterface
-        public void calledFromJS(String url) {
-            Log.d("АДРЕС С САЙТА: ", url);
-            VideoLink.videoLink = url;
-
-            Intent myIntent = new Intent(VideoLink.mainActivity, VideoActivity.class);
-            VideoLink.mainActivity.startActivity(myIntent);
-        }
+        webView.loadUrl("javascript:(function(){setTimeout(() => { let url = document.querySelector('.download-icon').href; VideoOpener.calledFromJS(url) }, 10000) })()");
     }
 
 }
