--- app/src/main/java/ru/ramazan/tubesleeper/WebAppInterface.java
@@ -1,22 +1,0 @@
-package ru.ramazan.tubesleeper;
-
-
-import android.content.Context;
-import android.webkit.JavascriptInterface;
-import android.widget.Toast;
-
-public class WebAppInterface {
-    Context mContext;
-
-    /** Instantiate the interface and set the context */
-    WebAppInterface(Context c) {
-        mContext = c;
-    }
-
-    /** Show a toast from the web page */
-    @JavascriptInterface
-    public void showToast(String toast) {
-        Toast.makeText(mContext, toast, Toast.LENGTH_SHORT).show();
-    }
-}
-
